% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-chpass.R
\name{module-chpass}
\alias{module-chpass}
\alias{chpass_ui}
\alias{chpass_server}
\title{Change password module}
\usage{
chpass_ui(id, tag_img = NULL, status = "primary", lan = NULL)

chpass_server(
  input,
  output,
  session,
  update_credentials,
  validate_pwd = NULL,
  use_token = FALSE,
  lan = NULL
)
}
\arguments{
\item{id}{Module's id.}

\item{tag_img}{A \code{tags$img} to be displayed on the authentication module.}

\item{status}{Bootstrap status to use for the panel and the button.
Valid status are: \code{"default"}, \code{"primary"}, \code{"success"},
\code{"warning"}, \code{"danger"}.}

\item{lan}{An language object. Should not be used directly.}

\item{input, output, session}{Standard Shiny server arguments.}

\item{update_credentials}{A \code{function} to perform an action when changing password is successful.
Two arguments will be passed to the function: \code{user} (username) and \code{password}
(the new password). Must return a list with at least a slot \code{result} with \code{TRUE}
or \code{FALSE}, according if the update has been successful.}

\item{validate_pwd}{A \code{function} to validate the password enter by the user.
Default is to check for the password to have at least one number, one lowercase,
one uppercase and be of length 6 at least.}

\item{use_token}{Add a token in the URL to check authentication. Should not be used directly.}
}
\description{
Change password module
}
